modularize layout code
