+++ src/library/masonry/mod.rs
@@ -1,0 +1,5 @@
+//! Justified-row masonry layout pieces for the photos grid.
+
+pub mod layout;
+pub mod load;
+pub mod quality;
