+++ src/library/thumbnail_cache.rs
@@ -290,16 +290,9 @@ impl ThumbnailCache {
         .map_err(|err| err.to_string())?;
 
         if let Some(bytes) = from_disk {
-            let byte_len = bytes.len();
             let texture = decode_to_scaled_texture(bytes, target_dim_for_bucket(size))
                 .await
                 .map_err(|err| err.to_string())?;
-            log::debug!(
-                "Thumbnail disk-cache hit for {:?} {}: {} bytes",
-                size,
-                asset_id,
-                byte_len
-            );
             self.memory.lock().insert(key.to_string(), texture.clone());
             return Ok(texture);
         }
@@ -307,10 +300,7 @@ impl ThumbnailCache {
         if is_cancelled() {
             return Err("cancelled".to_string());
         }
-        let fetch_started = std::time::Instant::now();
         let bytes = self.api_client.fetch_thumbnail(asset_id, size).await?;
-        let fetch_ms = fetch_started.elapsed().as_millis();
-        let byte_len = bytes.len();
         let cache_dir = self.cache_dir.clone();
         let cache_file_for_write = cache_file.clone();
         let bytes_for_write = bytes.clone();
@@ -322,14 +312,6 @@ impl ThumbnailCache {
         let texture = decode_to_scaled_texture(bytes, target_dim_for_bucket(size))
             .await
             .map_err(|err| err.to_string())?;
-        log::debug!(
-            "Thumbnail fetched from server for {:?} {}: {} bytes in {}ms",
-            size,
-            asset_id,
-            byte_len,
-            fetch_ms,
-        );
-
         self.memory.lock().insert(key.to_string(), texture.clone());
         Ok(texture)
     }
@@ -396,12 +378,6 @@ impl ThumbnailCache {
         .map_err(|err| err.to_string())?;
 
         if let Some(texture) = from_disk {
-            log::debug!(
-                "Local thumbnail disk-cache hit for {} ({}x{})",
-                path.display(),
-                texture.width(),
-                texture.height(),
-            );
             self.memory.lock().insert(key.to_string(), texture.clone());
             return Ok(texture);
         }
